auth/requirements.txt
ansible==8.6.1
ansible-core==2.15.6
cffi==1.16.0
cryptography==41.0.5
grpcio==1.59.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
protobuf==4.25.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
resolvelib==1.0.1
